+++ headers/constants.h
@@ -3,19 +3,19 @@
 
 enum color { YELLOW, RED };
 
-extern const unsigned N_ROWS;
-extern const unsigned N_COLUMNS;
-extern const unsigned NO_TOKEN;
-extern const unsigned YELLOW_TOKEN;
-extern const unsigned RED_TOKEN;
-extern const unsigned NO_ERROR;
-extern const unsigned COLUMN_FULL;
-extern const unsigned COLUMN_NUMBER_ERROR;
-extern const unsigned GAME_CONTINUE;
-extern const unsigned PLAYER_1_WON;
-extern const unsigned PLAYER_2_WON;
-extern const unsigned TEST_PASSED;
-extern const unsigned TEST_FAILED;
-extern const unsigned NO_WINNING_MOVE;
+extern const unsigned short N_ROWS;
+extern const unsigned short N_COLUMNS;
+extern const unsigned short NO_TOKEN;
+extern const unsigned short YELLOW_TOKEN;
+extern const unsigned short RED_TOKEN;
+extern const unsigned short NO_ERROR;
+extern const unsigned short COLUMN_FULL;
+extern const unsigned short COLUMN_NUMBER_ERROR;
+extern const unsigned short GAME_CONTINUE;
+extern const unsigned short PLAYER_1_WON;
+extern const unsigned short PLAYER_2_WON;
+extern const unsigned short TEST_PASSED;
+extern const unsigned short TEST_FAILED;
+extern const unsigned short NO_WINNING_MOVE;
 
 #endif
