+++ src/computer_game.c
@@ -1,10 +1,10 @@
 #include "computer_game.h"
 
-void play_with_computer(unsigned grid[6][7]) {
+void play_with_computer(unsigned short grid[6][7]) {
     /*
     Boucle infinie qui définie le comportement du jeu avec 2 joueurs humain.
     */
-    unsigned game_status, round_number = 1;
+    unsigned short game_status, round_number = 1;
 
     do {
         printf("\nTOUR %d\n", round_number);
